+++ packages/stacks-svelte/src/components/Checkbox/Checkbox.svelte
@@ -4,7 +4,9 @@
 </script>
 
 <script lang="ts">
+    import clsx from "clsx";
     import Label from "../Label/Label.svelte";
+    import type { ClassValue } from "svelte/elements";
 
     export type Props = {
         /**
@@ -60,7 +62,7 @@
         /**
          * Additional CSS classes
          */
-        class?: string;
+        class?: ClassValue;
 
         /**
          * Change event handler
@@ -99,26 +101,19 @@
     });
 
     const getClasses = (
-        className: string,
+        className: ClassValue,
         checkmark: boolean,
         state: State
     ) => {
         const base = "s-checkbox";
-        let classes = base;
-
-        if (className) {
-            classes += " " + className;
-        }
-
+        const classes: string[] = [];
         if (checkmark) {
-            classes += ` ${base}__checkmark`;
+            classes.push(`${base}__checkmark`);
         }
-
         if (state) {
-            classes += ` has-${state}`;
+            classes.push(`has-${state}`);
         }
-
-        return classes;
+        return clsx(base, className, classes);
     };
 
     const classes = $derived(getClasses(className, checkmark, state));