+++ packages/stacks-svelte/src/components/Popover/PopoverCloseButton.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+    import clsx from "clsx";
+    import type { ClassValue } from "svelte/elements";
     import Icon from "../Icon/Icon.svelte";
     import { IconCross } from "@stackoverflow/stacks-icons/icons";
     import { usePopoverContext } from "./Popover.svelte";
@@ -11,7 +13,7 @@
         /**
          * Additional CSS classes added to the element
          */
-        class?: string;
+        class?: ClassValue;
         /**
          * Callback fired when the close button is clicked
          */
@@ -30,7 +32,7 @@
 
 <button
     aria-label={label}
-    class={`s-popover--close s-btn s-btn__tonal ps-absolute${className ? ` ${className}` : ""}`}
+    class={clsx("s-popover--close s-btn s-btn__tonal ps-absolute", className)}
     type="button"
     onclick={handleClick}
 >